extract metadata from data_manager
