+++ src/meta_def/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "meta_def"
+version = "0.1.0"
+authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+sql_model = { path = "../sql_model" }
