+++ src/plan/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 ast = { path = "../ast" }
 constraints = { path = "../constraints" }
 data_manager = { path = "../data_manager" }
+meta_def = { path = "../meta_def" }
 protocol = { path = "../protocol" }
 sql_model = { path = "../sql_model" }
